refactor: create service file for fetch and remove libs folder in app
